1695 Move Case Info View to Separate Folder
