+++ web-client/src/views/CaseDetail/PartyInformation.jsx
